Changed parser caching to use the FullPath metadata instead of ItemSpec because depending on how you create your ItemGroup in MSBuild the ItemSpec can vary which will cause cache misses (and failures).
Added a default .targets file that can be imported into a project and provide all the necessary item creations and tasks. 

Added an AfterBuild step to the CodeGenerator that will copy the required assemblies and .targets file to the MSBuildExtensions folder. (Defaults to c:\program files\msbuild\castle.tools.codegenerator)