Added support for updating existing jobs and for listing the names of all existing jobs.
Added a new create job conflict action to update jobs instead of replacing them.  This is useful because it preserves execution history.  It is then quite straightforward to implement a routine on application startup that can purge obsolete scheduled jobs from the job store and update all existing ones so that version upgrades are seamless.

Renamed several properties to clearly document that UTC is being assumed in many places.  Added extra validations and unit tests to ensure that all dates are flagged with the UTC DateTimeKind so as to avoid any nasty surprises.

Merged FirstTime, JobSucceeded and JobFailed trigger schedule conditions.  Instead the PeriodicTrigger maintains its own IsFirstTime flag and all Triggers receive the most recent JobExecutionDetails from which to make decisions based on job history.

Redefined the Pending JobState so as to be more general than just the first time a job is added.  This was needed to support the new UpdateJob API since we'd need to somehow know to ask the trigger to reschedule itself.  Consequently the scheduler doesn't maintain any notion of whether a job is being triggered for the first time or not because it ceases to be meaningful when updates can be involved.

Refactored the AdoNetJobStore into a PersistentJobStore and an IJobStoreDao.  This additional layer of abstraction should make it quite easy to implement alternate Db-based backing stores just by implementing the DAO interface.  A candidate use-case would be an implementation based on ActiveRecord.

Added more tests.  Caught a few bugs related to Version number incrementation.