Added the FlexBridge Flash components including some common utilities and an Inversion of Control container with Constructor Dependency Injection.  These components originated in another project of mine called Odin and I've decided to refactor and contribute them to the general community.
Added some unit tests and reworked the test suite a bit.

Note: This checkin will break the build momentarily.  I'm committing it now so that I can safely move the code over to my laptop and finish it up there.  I wouldn't do this normally but it's very unlikely that anyone will be impacted right now.